pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llamaclaw"
version = "0.1.0"
description = "Lightweight Telegram chatbot and research assistant for Ollama."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115,<1.0",
  "httpx>=0.28,<1.0",
  "pydantic>=2.10,<3.0",
  "pydantic-settings>=2.6,<3.0",
  "python-telegram-bot>=22.0,<23.0",
  "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.25,<1.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
include = ["app*"]
